refactor the function handle
