+++ include/onnxruntime/core/graph/function.h
@@ -21,26 +21,11 @@ class Function {
  public:
   virtual ~Function() = default;
 
-#if !defined(ORT_MINIMAL_BUILD)
-  /** Gets the OpSchema for the Function. */
-  virtual const ONNX_NAMESPACE::OpSchema& OpSchema() const = 0;
-#endif
-
   /** Gets the Graph instance for the Function body subgraph. */
   virtual const onnxruntime::Graph& Body() const = 0;
 
   /** Gets the Mutable Graph instance for the Function body subgraph. */
   virtual onnxruntime::Graph& MutableBody() = 0;
-
-  virtual void Instantiated()= 0;
 };
 
-/** 
-Create a new Function instance.
-@param graph The graph containing the Function.
-@param nodes_to_fuse the IndexedSubGraph to use for the Function.
-*/
-std::unique_ptr<Function> MakeFunction(const onnxruntime::Graph& graph,
-                                       const IndexedSubGraph& nodes_to_fuse,
-                                       const logging::Logger& logger);
 }  // namespace onnxruntime
